requirements.txt
alpaca-py
krakenex
pykrakenapi
yfinance
pandas
pandas-ta-classic
ta
numpy
python-dotenv
pyyaml
requests
requests-cache
schedule
streamlit
pytest
flake8
optuna
scikit-learn
plotly